Refactor: moved runner code into a new GameManager class so all code in contained within objects
