+++ lib/require_helper.rb
@@ -4,3 +4,4 @@ require_relative 'player'
 require_relative 'turn_manager'
 require_relative 'computer_random'
 require_relative 'checker_criteria'
+require_relative 'game_manager.rb'
